requirements.txt
apache-airflow==2.7.3
kafka_python==2.0.2
yfinance==0.2.32
PyMySQL==1.1.0